pyproject.toml
[tool.poetry]
name = "orchestra-server"
version = "0.1.0"
description = ""
authors = ["jodafons <jodafons@lps.ufrj.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.2"
loguru = "^0.7.2"
flask-security = "^3.0.0"
passlib = "^1.7.4"
ptable = "^0.9.2"
cython = "^3.0.3"
flask-mail = "^0.9.1"
requests = "^2.31.0"
werkzeug = "^3.0.0"
jinja2 = "^3.1.2"
benedict = "^0.3.2"
prettytable = "^3.9.0"
psycopg2-binary = "^2.9.9"
sqlalchemy-utils = "^0.41.1"
ansi2html = "^1.8.0"
tqdm = "^4.66.1"
psutil = "^5.9.6"
pytest-order = "^1.1.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
orjson = "^3.9.9"
python-socketio = "^5.10.0"
gradio = "3.41.2"
expand-folders = "^1.0.0"
py-cpuinfo = "^9.0.0"
gputil = "^1.4.0"
wandb = "^0.15.12"
charset-normalizer = "^3.3.1"
tabulate = "^0.9.0"
nvsmi = "^0.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
